Code refactor with useWindowFocusRef.
